Rename : SoldOutStatus 클래스를 product.domain.type 패키지로 이동
